icicle-kit/image Added
Produce an SDcard capable of booting on the icicle-kit.
The partitions are given appropreate types and the bootloader is
packaged as a hss payload.

Pipewire is disabled since it is causing corectl to be spammed.

gdm is forced to use software rendering because hardware acceleration
appears to be broken on this board.

Co-authored-by: Michael Drake <michael.drake@codethink.co.uk>