icicle-kit/uboot Generate a HSS Payload
the Icicle-kit requires uboot to be packaged as a HSS payload in
order to be booted from.
This uses a pre-compiled x86 binary 'hss-payload-generator',
which we can get away with using in the riscv environment on account
of it being statically linked.
Bootargs are based on the ones used in the buildroot-BSP.