gnomeos: Add an installer prototype for the live image
This add a prototype of installer. This was written in Python just to
accelerate development, but it should changed from now to something a
bit better. All the parts are in, so now we can work on testing and
improving the experience.

For now the installer will on purpose refuse to install if the disk
has some data. The user will have to open Disks, then select the disk
to clean up, then "format", and choose "no partitioning". Then the
installer will allow installation to happen. This is not a problem for
VMs like Boxes, because the initial disk is usually empty. Once we
have a proper interface that is clear enough that we are removing
data, we can remove that limitation.

The installer application will only appear in live, that is if kernel
parameter `root=live:gnomeos` is set. And polkit will also only allow
installation from `wheel` users if we are in live mode.

There are 2 modes of installation, either we install on the disk the
old fashion way, then after reboot gnome-initial-setup will kick in
again for the installed OS. That is called for now "OEM installation"
on the installer.

Or by default, it will just write running system into the disk. After
reboot the data will still be available.

If a TPM 2.0 is present, encryption will happen automatically.  This
will use PCR 11 signatures, and pcrlock. PCR 7 will not be statically
enrolled, but pcrlock can take care of that. We initially do not lock
the UEFI firmare configuration, because the boot entry will be created
on first reboot. But after first reboot, it will lock itself
automatically. Also secure boot will be locked automatically once
after the users enrolls secure boot keys.

If disk encryption happens, then the user will get a recovery password
printed in the installer.