[Endless] linux-pam: Enable usergroups for private user groups, the UMASK 0002
With every user having their own group, the "historical" umask of 0022
is excessively restrictive. For example, create an user account "user1"
on GNOME OS current version. The user will have username "user1" in
primary group "user1" but umask is 0022. The main disadvantage is that
it prevents setting up shared directories for common access.

To enable this type of shared directories, add the pam_umask module
configured with usergroups=true so that the default umask becomes 0002
(instead of 0022). Then, new files in shared directories can be read &
written by the users in the same group.

AC:
- Command "umask" should output 0002.
- 1. Have two users: user1 and user2 on the system.
  2. Add user2 into user1's group "user1" by command:
     "sudo usermod -aG user1 user2".
  3. User1 creates a folder by command: "mkdir -p /tmp/test". Then,
     create a dummy file into it by "touch /tmp/test/dummy.txt".
  4. User1 logs out. And, login with user2. User2 should be able to
     modify "/tmp/test/dummy.txt".

https://phabricator.endlessm.com/T31847
https://phabricator.endlessm.com/T35802
https://github.com/endlessm/eos-build-meta/pull/2#discussion_r2050472931