[Endless] ostree: deploy: Create fake symlinks when on FAT filesystems
When trying to deploy on a vfat partitions, such as the EFI ESP, we
fail to create the loader symlink, and deployment fails.

We'd like to use sd-boot for PAYG deployment, which only works with the
ESP, so we need a way to deploy without using symlinks.

For a quick and dirty minimally invasive fix, we've made deployment
create a bogus symlink - a text file named loader.sln that contains
the name of the real loader dir.

We've modified systemd-boot to read these bogus symlinks as well.

AC:
1. Fake symlinks ("/boot/loader.sln") should exits on EOS PAYG
   deployment system and has content "loader.0" or "loader.1". Non-PAYG
   system should not have the fake symlinks ("/boot/loader.sln").
2. EOS PAYG deployed system should boot into desktop environment.

https://phabricator.endlessm.com/T27040
https://phabricator.endlessm.com/T35831