[Endless] systemd: dissect-image: Allow non-root to read the ESP in /boot
On PAYG images we mount the ESP to /boot and install the boot loader
configuration and entries there, as this is a requirement of
systemd-boot. The GPT auto generator finds the ESP and generates an
automount unit for it at /boot in a way that makes it only accessible
and readable by root. OSTree needs to read the contents of
/boot/loader/entries for non-root `ostree admin status`, and probably
other similar actions.

This commit changes the umask used in boot.mount to allow for group and
others to read files and access directories in the ESP mounted to /boot.

AC: Non-root can execute `ostree admin status` directly.

https://phabricator.endlessm.com/T29520
https://phabricator.endlessm.com/T35070