Add `eos/repo.bst` element
EOS7 now defines its own filesystem instead of reusing the one from
GNOME OS.

The proof-of-concept "signed boot" image is gone.  This is necessary:
the GNOME OS elements define a complete image pipeline, which includes
signing the image and generating the initramfs, and does not support
OSTree.

EOS deploys to OStree and there is existing tooling outside of
BuildStream for signing and image creation.

Some of the files removed in this commit may need to reappear later
with modifications when we implement secure boot.

The new `eos/repo.bst` element is the final stage of the BuildStream
build, producing an OSTree repo that holds the filesystem tree.