Sign and push os trees in CI
The `ostree-push` command is made available via `apt install`, we
don't need the corresponding BuildStream element.

Note that the ostree-receive running on the internal Endless server
is an internal version that formed the basis of what's in the public
ostree-push project. This necessitates passing an extra option to
ostree-push to run `ostree-receive` instead of a versioned binary
`ostree-receive-1`.

Also, the old ostree-receive doesn't cope well with the absolute repo
path parsed from an ssh:// URL. With the scp style URL, it's clearly a
relative path. When ostree-receive is updated on the server, we can go
back to the ssh:// URL.