Don't include locally generated signing key in the image
The real Endless signing key is now included via the `eos-keyring.bst`
element, added for https://github.com/endlessm/eos-build-meta/issues/41.

Including a locally generated signing key is a problem for CI -- it
it generates a randomly different signing key on each pipeline, so
each pipeline produces a slightly different filesystem artifact that
is never reused. See:
https://github.com/endlessm/eos-build-meta/issues/56.

This was anyway only added as a developer convenience for local builds.