Add Shell extensions
This adds the GNOME Shell extensions used by Endless OS.

Overview Background extension is incompatible with the recent GNOME Shell
we have. So with this commit, we do install it but not enable it.

Also, if you're wondering, in `gnome-shell-extensions.bst`, we're using
a `manual` kind instead of `stack` kind because we need to run
install-commands to setup all the dconf settings/db for all the
extensions.

Fixes #7.

Co-authored-by: Adrien Plazas <adrien.plazas@codethink.co.uk>