Update systemd-tmpfiles config to rename 'wheel' group
Both systemd and Freedesktop SDK work on the assumption that the
`wheel` group gives superpowers. In EOS6 the group is named
`sudo` and we re-use that name in EOS7 to avoid a migration step.

This means patching the tmpfiles config, which we do as an
integration step so we can re-use the elements Freedesktop SDK
that integrate systemd and cups.