Add the prebuilt Shim binaries
It's possible to build these from source, but we currently have
no way to sign them.

For the fallback (`fbx64.efi` and mokmanager (`mmx64.efi`) binaries,
we can build from source, and sign with the Endless cert once
<https://github.com/endlessm/eos-build-meta/issues/89> is completed.

For the shim bootloader (`shimx64.efi`) we need Microsoft to sign
the binary so UEFI firmwares accept the signature directly, so in
the short term we aren't going to build it from source.