Update dracut-ng to 108
This brings a more reliable way to parse shared object dependencies.
For details see:
<https://github.com/dracut-ng/dracut-ng/commit/19b5faad97971a6d906740de28cfe366a7fae687>.

Previously our eos7 initramfs was failing because dracut didn't include
`libkmod.so.2` in the initramfs. We didn't notice because the code to
include libkmod.so.2 in 107 was hacky and didn't detect failure at all.
The new implementation logs an ERROR: message if expected files aren't
found.