Patch Flatpak to avoid sandboxing extra_data scripts
This is due to a limitation in eos-image-builder. The
`hooks/image/50-flatpak.chroot` image hook runs inside of `chroot()`,
and uses libflatpak calls to install apps and runtimes. When run on
an EOS7 buildroot, it fails with an error from `bwrap`, which is unable to
create user namespaces inside the chroot, while trying to run an
extra_data script. This worked with an EOS6 buildroot.

The limitation on creating user namespaces inside a chroot is hardcoded
in Linux itself (see `create_user_ns()` in `kernel/user_namespaces.c`).

So the change must be that Flatpak previously ran the extra_data script
without asking `bwrap` to create a user namespace.

See https://github.com/endlessm/eos-build-meta/issues/102.