Use 'snakeoil' signing elements for GRUB and kernel
This gives one way of testing EOS7 with UEFI Secure Boot. It's quick
and relatively simple and means that all builds have signatures.

Nobody should trust this signature except when testing UEFI Secure Boot.
In this case you'd manually enrol the vendor key as a Machine Owner Key
in the target machine's firmware.

This doesn't exercise the eos-sb-signer codepaths that will be used for
official builds.