Mount EFI at /efi instead of /boot
In GNOME OS, the `/boot` directory on the root partition is empty,
and it makes sense to mount the EFI system partition there.

In Endless OS, as we use OSTree to deploy the system, there are
files in `/boot` on the root partition that are not in the ESP.

So let's mount the ESP at /efi. This allows systemd-boot's `bootctl`
command to find it.