Move module signing into kernel build
This behaves differently depending on the `signed_boot` option.

  * `snakeoil`: The pregenerated snakeoil signing keys are used, as in
    previous commits. All signing now happens during the kernel build
    process instead of separately.
  * `endless`: A unique signing key is generated, used to sign
    the modules and then thrown away.

Using an ephemeral signing key means that official builds of linux.bst
are not bit-for-bit reproducible, and each CI build of main produces
many MBs of new `.ko.xz` files that may only differ in the signature
used. It means that all modules must be known and signed as part of
the `linux.bst` build process. And it means the risk of a given module
key being compromised is very low indeed, as for successful builds, the
key is never stored anywhere outside the transient build sandbox.

See `doc/overview/signing.md` for more details on signing.