Use systemd v256 with Endless patches
At least one patch to systemd-boot is essential for PAYG
support. (Which enables reading the `loaders/` folder via
a symlink in a text file -- necessary as OSTree requires
some sort of symlink support to deploy the boot config,
even on the ESP which is a FAT filesystem).

See: https://github.com/endlessm/eos-build-meta/issues/161