ostree: Reduce divergence with upstream
This introduces elements/include/ostree-source.yml back to reduce the
divergence with upstream and to let it be updated with
utils/update-elements.sh, limiting the risks of missing changes when
updating.

This also removes dependency on components/gtk-doc.bst as it's not in
the upstream version and we don't need it.