Create an empty /etc/machine-id file
This is needed for the ISO image to work. If the file exists and is
empty, systemd can generate a temporary /run/machine-id file and
bind-mount it over /etc/machine-id. If the file doesn't exist,
systemd can't do that, and instead prints a dire warning predicting
an impending boot failure.

Fixes https://github.com/endlessm/eos-build-meta/issues/95