Add a symlink /usr/local -> /var/local
This is required at least for the live ISO, as the
`eos-transient-setup` service writes to this directory.

It's also recommended by OSTree in the "System layout" docs:
https://ostreedev.github.io/ostree/adapting-existing/#system-layout

While OSTree suggests `/usr/local -> /var/usrlocal`, we use the
target name used in eos6 which was `/var/local`.

Part of https://github.com/endlessm/eos-build-meta/issues/14